updated schemas. added get/post to questions
