+++ public/views/single.html
@@ -7,12 +7,12 @@
     <script src="js/jquery.min.js"></script>
     <!-- Custom Theme files -->
     <!--menu-->
-    <script src="js/imagesPromptingHandler.js"></script>
-    <script src="js/app.js"></script>
-    <script src="js/sweetalert.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="css/sweetalert.css">
-    <link href="css/styles.css" rel="stylesheet">
-    <!--//menu-->
+    <!-- <script src="js/imagesPromptingHandler.js"></script>-->
+     <script src="js/app.js"></script>
+     <script src="js/sweetalert.min.js"></script>
+     <link rel="stylesheet" type="text/css" href="css/sweetalert.css">
+     <link href="css/styles.css" rel="stylesheet">
+     <!--//menu-->
     <!--theme-style-->
     <link href="css/style2.css" rel="stylesheet" type="text/css" media="all"/>
     <!--//theme-style-->
@@ -30,12 +30,14 @@
             src="https://code.angularjs.org/1.3.0-beta.5/angular.js">
     </script>
 
-    <script src="js/controllers/ListingCtrl.js"></script>
     <script src="js/services/ListingService.js"></script>
+    <script src="js/services/QuestionService.js"></script>
+    <script src="js/controllers/ListingCtrl.js"></script>
+    <script src="js/controllers/QuestionCtrl.js"></script>
 
 
 </head>
-<body ng-controller="ListingController">
+<body ng-controller="ListingCtrl">
 <!--header-->
 <div class="navigation">
     <div class="container-fluid">
